feat(settings panel): adds echo client and handler

+++ shell/settings-panel/src/modules/brightness/service.rs
@@ -28,12 +28,12 @@ impl BrightnessService {
             }
         };
 
-        Ok(brightness / 254 * 100)
+        Ok((brightness as f32 / 254. * 100.) as u8)
     }
 
     pub async fn set_brightness_value(value: u8) -> Result<()> {
         let task = "set_brightness_value";
-        match Display::set_brightness_percentage(value / 100 * 254).await {
+        match Display::set_brightness_percentage((value as f32 / 100. * 254.) as u8).await {
             Ok(v) => v,
             Err(e) => bail!(SettingsPanelError::new(
                 SettingsPanelErrorCodes::SetBrightnessError,
